refactor ContactForm component
