+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.5",
     "@testing-library/user-event": "^12.7.0",
+    "axios": "^0.21.1",
     "eslint-plugin-import": "^2.22.1",
     "final-form": "^4.20.1",
     "final-form-material-ui": "^0.3.0",
